backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
supabase==2.9.0
python-dotenv==1.0.1
serpapi==0.1.5
google-search-results==2.4.2
